Flatten folder structure
