+++ MANIFEST.in
@@ -1,4 +1,3 @@
 include LICENSE
 include README.md
 include holidays/fixtures/*
-include holidays/fixtures/*
